+++ classes/local/participant/entities/participant_interface.php
@@ -122,6 +122,13 @@ interface participant_interface {
     public function get_callsign();
 
     /**
+     * Returns pilot's fleet association
+     *
+     * @return string   The participant's fleet
+     */
+    public function get_fleet();
+
+    /**
      * Returns a participant's user profile field
      *
      * @param string $field     The name of the field